+++ src/main/kotlin/com/mahocommerce/maho/lsp/MahoSettings.kt
@@ -15,4 +15,5 @@ class MahoSettings : SimplePersistentStateComponent<MahoSettingsState>(MahoSetti
 
 class MahoSettingsState : BaseState() {
     var phpCommand by string("php")
+    var autoConfigureDatabase by property(true)
 }